22.gün tamamlandı

+++ app/src/app/pages/todo-list/todo-list.component.ts
@@ -1,7 +1,8 @@
 import { Component, OnInit } from '@angular/core';
-import { Todo } from 'src/app/models/todo';
-import { TodoService } from '../todo.service';
 
+import { TodoService } from '../todo.service';
+import sweetAlert from 'sweetalert2';
+import { Todo } from 'src/app/models/todo';
 @Component({
   selector: 'app-todo-list',
   templateUrl: './todo-list.component.html',
@@ -9,29 +10,46 @@ import { TodoService } from '../todo.service';
 })
 export class TodoListComponent implements OnInit {
   todoList: Todo[] = [];
+
   constructor(private todoService: TodoService) {}
   ngOnInit(): void {
     this.load();
   }
+
   load() {
     this.todoService.getAll().subscribe((x) => (this.todoList = x));
   }
+
   delete(id: number) {
-    this.todoService.delete(id).subscribe((x) => {
-      if (x) {
-        this.load();
-      } else {
-        alert('bir hata oluştu.');
-      }
-    });
+    sweetAlert
+      .fire({
+        title: 'Are you sure',
+        text: "you won't be able to revert this action",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonText: 'yes, delete it',
+      })
+      .then((result) => {
+        if (result.isConfirmed) {
+          this.todoService.delete(id).subscribe((x) => {
+            if (x) {
+              this.load();
+            } else {
+              alert('bir hata meydana geldi');
+            }
+          });
+        }
+      });
   }
 
   isCompleted(id: number) {
     this.todoService.isCompleted(id).subscribe((x) => {
+      console.log(x);
+
       if (x == true) {
         this.load();
       } else {
-        alert('bir hata meydana geldi.');
+        alert('bir hata meydana geldi');
       }
     });
   }
